--- app/src/main/java/com/benatt/passwordsmanager/data/db/PasswordsDatabase.java
@@ -1,20 +1,0 @@
-package com.benatt.passwordsmanager.data.db;
-
-import androidx.room.Database;
-import androidx.room.RoomDatabase;
-
-import com.benatt.passwordsmanager.data.models.passwords.PasswordDao;
-import com.benatt.passwordsmanager.data.models.passwords.model.Password;
-import com.benatt.passwordsmanager.data.models.user.UserDao;
-import com.benatt.passwordsmanager.data.models.user.model.User;
-
-import static com.benatt.passwordsmanager.utils.Constants.DB_VERSION;
-
-/**
- * @author bernard
- */
-@Database(entities = {User.class, Password.class}, version = DB_VERSION, exportSchema = false)
-public abstract class PasswordsDatabase extends RoomDatabase {
-    public abstract UserDao userDao();
-    public abstract PasswordDao passwordDao();
-}
