--- app/src/main/java/com/benatt/passwordsmanager/data/models/user/UserRepository.java
@@ -1,16 +1,0 @@
-package com.benatt.passwordsmanager.data.models.user;
-
-import com.benatt.passwordsmanager.data.models.Dao;
-import com.benatt.passwordsmanager.data.models.passwords.model.Password;
-
-import javax.inject.Inject;
-
-/**
- * @author bernard
- */
-public class UserRepository extends Dao<Password> {
-
-    @Inject
-    public UserRepository() {}
-
-}
